Renaming worker-store-in-memory to worker-store-mem
