+++ pom.xml
@@ -98,7 +98,7 @@
         <module>worker-example</module>
         <module>worker-archetype</module>
         <module>worker-store-s3</module>
-        <module>worker-store-in-memory</module>
+        <module>worker-store-mem</module>
     </modules>
 
     <scm>
