+++ worker-framework/docs/pages/en-us/apidocs/index.html
@@ -8,7 +8,7 @@
 <script type="text/javascript">
     targetPage = "" + window.location.search;
     if (targetPage != "" && targetPage != "undefined")
-        targetPage = targetPage.substring(1);
+        targetPage = targetPage.regexGroupSubstring(1);
     if (targetPage.indexOf(":") != -1 || (targetPage != "" && !validURL(targetPage)))
         targetPage = "undefined";
     function validURL(url) {
