--- worker-store-in-memory/src/main/resources/META-INF/services/com.hpe.caf.api.worker.DataStoreProvider
@@ -1,1 +1,0 @@
-com.hpe.caf.worker.datastore.inmemory.InMemoryDataStoreProvider
