+++ worker-store-mem/pom.xml
@@ -22,7 +22,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.github.workerframework</groupId>
-    <artifactId>worker-store-in-memory</artifactId>
+    <artifactId>worker-store-mem</artifactId>
     <version>1.5.0-SNAPSHOT</version>
 
     <parent>
