+++ worker-store-mem/src/main/resources/META-INF/services/com.hpe.caf.api.worker.DataStoreProvider
@@ -1,0 +1,1 @@
+com.hpe.caf.worker.datastore.mem.InMemoryDataStoreProvider
