+++ worker-store-mem/src/test/resources/testDataStore_path.txt
